Plugin/tests/com/n7space/capellatasteplugin/capellaintegration/datamodelelementinterpretationproviders/NumericValueElementInterpretationProviderTests.java
// N7 Space Sp. z o.o.
// www.n7space.com
// COPYRIGHT 2020

// Licensed under the terms of:
// Eclipse Public License available at http://www.eclipse.org/legal/epl-v10.html

package com.n7space.capellatasteplugin.capellaintegration.datamodelelementinterpretationproviders;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNull;
import static org.junit.Assert.assertTrue;

import java.util.ArrayDeque;
import java.util.LinkedList;
import java.util.List;

import org.junit.Test;
import org.polarsys.capella.core.data.information.datatype.NumericType;
import org.polarsys.capella.core.data.information.datatype.NumericTypeKind;
import org.polarsys.capella.core.data.information.datavalue.NumericValue;

import com.n7space.capellatasteplugin.capellaintegration.datamodelelementinterpretationproviders.DataModelElementInterpreter.DataModelElementInterpretationProvider;
import com.n7space.capellatasteplugin.capellaintegration.datamodelelementinterpretationproviders.testmocks.DataPackageMock;
import com.n7space.capellatasteplugin.capellaintegration.datamodelelementinterpretationproviders.testmocks.NumericTypeMock;
import com.n7space.capellatasteplugin.modelling.data.DataModel;
import com.n7space.capellatasteplugin.modelling.data.DataModelElement;
import com.n7space.capellatasteplugin.modelling.data.DataPackage;
import com.n7space.capellatasteplugin.modelling.data.DataType;
import com.n7space.capellatasteplugin.utils.Issue;

public class NumericValueElementInterpretationProviderTests {

	public NumericValue createIntegerValue(final String name, final Integer value) {
		final DataPackageMock pkgMock = new DataPackageMock("DummyPkg", "DummyPkg_id");
		final NumericType typeMock = new NumericTypeMock(pkgMock, name + "_type", name + "_type_id",
				NumericTypeKind.INTEGER);
		final NumericValue mock = new NumericTypeMock.NumericValueMock(pkgMock, name, name + "_id", typeMock,
				value.toString());
		return mock;
	}

	@Test
	public void testInterpretModelElement() {
		final DataModel model = new DataModel();
		final DataPackage pkg = new DataPackage("x", "x");
		final NumericValue mock = createIntegerValue("IntegerValue", Integer.valueOf(1969));
		final List<Issue> issues = new LinkedList<Issue>();

		final DataModelElementInterpretationProvider provider = new NumericValueElementInterpretationProvider();
		final DataModelElement element = provider.interpretModelElement(model, pkg, new ArrayDeque<DataModelElement>(),
				mock, issues);
		assertTrue(element instanceof DataType.DataTypeValue);
		assertEquals(element.name, "IntegerValue");
		assertEquals(((DataType.DataTypeValue) element).type.id, "IntegerValue_type_id");
		assertEquals(((DataType.DataTypeValue) element).value, "1969");
	}

	@Test
	public void testInterpretModelElement_untyped() {
		final DataModel model = new DataModel();
		final DataPackage pkg = new DataPackage("x", "x");
		final DataPackageMock pkgMock = new DataPackageMock("DummyPkg", "DummyPkg_id");
		final NumericValue mock = new NumericTypeMock.NumericValueMock(pkgMock, "IntegerValue", "IntegerValue" + "_id",
				null, "1969");
		final List<Issue> issues = new LinkedList<Issue>();

		final DataModelElementInterpretationProvider provider = new NumericValueElementInterpretationProvider();
		final DataModelElement element = provider.interpretModelElement(model, pkg, new ArrayDeque<DataModelElement>(),
				mock, issues);
		assertNull(element);
		assertEquals(issues.size(), 1);
	}

}
